requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pdfplumber==0.10.3
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.13
qdrant-client==1.7.0
pydantic==2.5.3
python-dotenv==1.0.1